Added KiCad Schematic, Software Updates
Added the KiCad schematic showing the circuit layout of the box.  Compiled a list of parts needed for the box in an Excel sheet.  Made changes to the software - changed some X1 pins to now be Input Only, and adjusted the test creation software to reflect this change.
Still TODO:
Fix the Pin Input system - Relay_Test has pins having faulty input.  Adjust the program to show which pins have wrong input.